Fix DiscPlayer savestate framebuffer length width
Encode framebuffer length as u32 to match decode. Strengthen the player
save-state test with non-zero corner pixels so length/shift bugs cannot
hide behind a zero framebuffer.

+++ crates/playdia-core/src/video.rs
@@ -261,7 +261,7 @@ impl VideoDecoder {
             out.extend_from_slice(&(frame.len() as u32).to_le_bytes());
             out.extend_from_slice(frame);
         }
-        out.extend_from_slice(&self.framebuffer.len().to_le_bytes());
+        out.extend_from_slice(&(self.framebuffer.len() as u32).to_le_bytes());
         for px in &self.framebuffer {
             out.extend_from_slice(&px.to_le_bytes());
         }
